requirements.txt
fastapi
uvicorn
pydantic
python-multipart
google-generativeai
chromadb
PyMuPDF
tqdm
